third_party/blink/web_tests/fast/events/onchange-setvalue-expected.txt

PASS events.length is 2
PASS events[0] is 'change e'
PASS events[1] is 'change es'
PASS events.length is 3
PASS events[0] is 'change a'
PASS events[1] is 'change ab'
PASS events[2] is 'change abc'
PASS events.length is 2
PASS events[0] is 'change e'
PASS events[1] is 'change es'
PASS events.length is 2
PASS events[0] is 'change e'
PASS events[1] is 'change es'
PASS successfullyParsed is true

TEST COMPLETE
